+++ src/events/client/disconnect.js
@@ -1,5 +1,7 @@
 const logger = require('../../services/logger');
-const { embedOptions, systemOptions } = require('../../config');
+const config = require('config');
+const embedOptions = config.get('embedOptions');
+const systemOptions = config.get('systemOptions');
 const { EmbedBuilder } = require('discord.js');
 
 module.exports = {