+++ src/utils/factory/createPlayer.js
@@ -1,6 +1,5 @@
 const logger = require('../../services/logger');
-const { Player, onBeforeCreateStream } = require('discord-player');
-const { stream } = require('yt-stream');
+const { Player } = require('discord-player');
 
 exports.createPlayer = async (client) => {
     try {
@@ -23,6 +22,7 @@ exports.createPlayer = async (client) => {
         // primarily to be able to use it in broadcastEval and other sharding methods
         global.player = player;
 
+        /* Seems to not be needed with @distube/ytdl-core, long duration videos work fine, keeping commented for now
         onBeforeCreateStream(async (track) => {
             if (track.source === 'youtube') {
                 return (
@@ -37,6 +37,7 @@ exports.createPlayer = async (client) => {
 
             return null;
         });
+        */
 
         await player.extractors.loadDefault();
         logger.trace(`[Shard ${client.shard.ids[0]}] discord-player loaded dependencies:\n${player.scanDeps()}`);